+++ src/onegov.election_day/onegov/election_day/tests/collections/test_notification_collection.py
@@ -78,6 +78,8 @@ def test_notification_collection(session):
 
         # Add a email, SMS notification and webhook
         request = DummyRequest(session=session)
+        request.app.mail_sender = 'info@onegov.ch'
+        request.app.locales = ['de_CH', 'it_CH', 'fr_CH', 'rm_CH']
         request.app.principal.email_notification = True
         request.app.principal.sms_notification = 'http://example.com'
         request.app.principal.webhooks = {'http://abc.com/1': None}